# imswitch/_data/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/SettingsController/setDetectorExposureTime": {
      "get": {
        "summary": "Setdetectorexposuretime",
        "description": "Sets the exposure time for the specified detector.",
        "operationId": "setDetectorExposureTime_SettingsController_setDetectorExposureTime_get",
        "parameters": [
          {
            "name": "detectorName",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Detectorname" }
          },
          {
            "name": "exposureTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1,
              "title": "Exposuretime"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/SettingsController/setDetectorGain": {
      "get": {
        "summary": "Setdetectorgain",
        "description": "Sets the gain for the specified detector.",
        "operationId": "setDetectorGain_SettingsController_setDetectorGain_get",
        "parameters": [
          {
            "name": "detectorName",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Detectorname" }
          },
          {
            "name": "gain",
            "in": "query",
            "required": false,
            "schema": { "type": "number", "default": 0, "title": "Gain" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/SettingsController/setDetectorParameter": {
      "get": {
        "summary": "Setdetectorparameter",
        "description": "Sets the specified detector-specific parameter to the specified\nvalue.",
        "operationId": "setDetectorParameter_SettingsController_setDetectorParameter_get",
        "parameters": [
          {
            "name": "detectorName",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Detectorname" }
          },
          {
            "name": "parameterName",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Parametername" }
          },
          {
            "name": "value",
            "in": "query",
            "required": true,
            "schema": { "title": "Value" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/RecordingController/snapNumpyToFastAPI": {
      "get": {
        "summary": "Snapnumpytofastapi",
        "description": "Taking a snap and return it as a FastAPI Response object.\ndetectorName: the name of the detector to take the snap from. If None, take the snap from the first detector.\nresizeFactor: the factor by which to resize the image. If <1, the image will be downscaled, if >1, nothing will happen.",
        "operationId": "snapNumpyToFastAPI_RecordingController_snapNumpyToFastAPI_get",
        "parameters": [
          {
            "name": "detectorName",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Detectorname" }
          },
          {
            "name": "resizeFactor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1,
              "title": "Resizefactor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}